Fix bug where pasting inserts ANSI escape codes
Remapping <Esc> to <nop> was done in order to retrain myself to use the
new jk binding I had added to exit insert mode. Now that I have learned
this behavior, it's fine to give its original behavior back. It causes
issues with how certain signals are sent (such as pasting, which sends
escape codes 200 and 201 surrounding the text.)

Also disabled the <nop> on <C-c> as there's no longer a need.